+++ R/regr_smape.R
@@ -10,7 +10,7 @@
 #' @templateVar mid smape
 #' @template regr_template
 #'
-#' @note
+#' @details
 #' This measure is undefined if if any \eqn{|t| + |r|} is \eqn{0}.
 #'
 #' @inheritParams regr_params
@@ -19,8 +19,9 @@
 smape = function(truth, response, na_value = NaN, ...) {
   assert_regr(truth, response = response, na_value = na_value)
   denom = abs(truth) + abs(response)
-  if (any(denom < TOL))
+  if (any(denom < TOL)) {
     return(na_value)
+  }
   2 * mean(ae(truth, response) / denom)
 }